% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_volatility_arbitrage.R
\name{compute_index_weights}
\alias{compute_index_weights}
\title{Print "Compute Index Weights"}
\usage{
compute_index_weights(marketCapitalizations)
}
\arguments{
\item{marketCapitalizations}{: a vector of market capitalizations of the index's constituents}

\item{shareFloats}{: a numeric vector of shares}
}
\value{
weights : a vector of weights for the index
}
\description{
This function computes the weights of the index constituents based on the market capitalization based on the floats
}
