% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_volatility_arbitrage.R
\name{compute_market_capitalization}
\alias{compute_market_capitalization}
\title{Compute Market Capitalization}
\usage{
compute_market_capitalization(shareCount, sharePrice)
}
\arguments{
\item{shareCount}{: count of shares}

\item{sharePrice}{: price per share}
}
\value{
Returns a market capitalization or a vector of capitalizations
}
\description{
This function returns market capitalizations for a vector of stocks
}
