% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_volatility_arbitrage.R
\name{compute_index_volatility}
\alias{compute_index_volatility}
\title{Print "Compute Index Weights"}
\usage{
compute_index_volatility(indexWeights, constituentVolatilities, indexCovariance)
}
\arguments{
\item{indexWeights}{: a vector of market capitalizations of the index's constituents}

\item{constituentVolatilities}{}

\item{indexCovariance}{}
}
\value{
indexVolatility : a volatility for the index
}
\description{
This function computes the weights of the index constituents based on the market capitalization based on the floats
}
